Time: 2 ms (83.84%), Space: 149.1 MB (73.29%) - LeetHub

+++ 3994-minimum-adjacent-swaps-to-partition-array/3994-minimum-adjacent-swaps-to-partition-array.cpp
@@ -26,7 +26,7 @@ public:
             }
         }
 
-        return swaps ;
+        return (int)swaps ;
         
     }
 };
